requirements.txt
Django==4.1.7
django-svelte==0.1.6
djangorestframework==3.14.0
drf-spectacular==0.26.0
